feature: update product controller - implement logic to add user on product creation - implement validation while updating and deleting product - add docker file

+++ src/contoller/product.controller.ts
@@ -9,8 +9,13 @@ import { type CreateProductInput, type DeleteProductInput, type UpdateProductInp
 import { createProduct, deleteProduct, findAndUpdateProduct, findProduct } from '../services/product.service';
 import ErrorResponse from '../utils/errorResponse';
 import multerInstance from '../utils/multerInstance';
+import { type User } from '../model/user.model';
+import type mongoose from 'mongoose';
+import { findUser } from '../services/user.service';
+import { gt, isEqual } from 'lodash';
 
 const upload = multerInstance.single('uploaded_file');
+type RequestUserType = User & { _id: mongoose.Types.ObjectId };
 
 export const uploadHandler = (req: Request, res: Response, next: NextFunction) => {
   upload(req, res, function (err) {
@@ -33,7 +38,14 @@ export const createProductHandler = async (
 ) => {
   try {
     const { name, category, description, feature, price } = req.body;
+    const { _id, email } = req.user as RequestUserType;
     const imagePath = req.file?.path;
+    const createdBy = _id;
+    const user = await findUser({ email });
+    if (isEmpty(user)) {
+      next(new ErrorResponse(MESSAGES.USER_NOT_FOUND, 400));
+      return;
+    }
     const existingProduct = await findProduct({ name });
     if (!isEmpty(existingProduct)) {
       next(new ErrorResponse(MESSAGES.PRODUCT_EXISTS, 400));
@@ -46,7 +58,10 @@ export const createProductHandler = async (
       name,
       price: parseInt(price),
       imagePath,
+      createdBy,
     });
+    user.products.push(newProduct.id);
+    await user.save();
     res.status(200).json({ data: newProduct });
   } catch (error) {
     logger.error(error);
@@ -63,8 +78,15 @@ export const updateProductHandler = async (
   next: NextFunction
 ) => {
   try {
+    const { _id } = req.user as RequestUserType;
     const { productId } = req.params;
     const existingProduct = await findProduct({ _id: productId });
+
+    if (!isEqual(existingProduct?.createdBy, _id)) {
+      next(new ErrorResponse(MESSAGES.NOT_AUTHORIZED, 403));
+      return;
+    }
+
     if (isEmpty(existingProduct)) {
       next(new ErrorResponse(MESSAGES.PRODUCT_DOESNOT_EXISTS, 400));
       return;
@@ -83,12 +105,29 @@ export const updateProductHandler = async (
 export const deleteProductHandler = async (req: Request<DeleteProductInput>, res: Response, next: NextFunction) => {
   try {
     const { productId } = req.params;
+    const { _id, email } = req.user as RequestUserType;
+
     const existingProduct = await findProduct({ _id: productId });
+    const user = await findUser({ email });
+    const productIndexInUser = user?.products.findIndex((product) =>
+      isEqual((product as mongoose.Types.ObjectId).toString(), productId)
+    );
+
+    if (!isEqual(existingProduct?.createdBy, _id)) {
+      next(new ErrorResponse(MESSAGES.NOT_AUTHORIZED, 403));
+      return;
+    }
+
     if (isEmpty(existingProduct)) {
       next(new ErrorResponse(MESSAGES.PRODUCT_DOESNOT_EXISTS, 400));
       return;
     }
     await deleteProduct({ _id: productId });
+    if (gt(productIndexInUser, -1)) {
+      user?.products.splice(productIndexInUser as number, 1);
+      await user?.save();
+    }
+
     unlinkSync(existingProduct.imagePath);
     res.status(200).json({ data: MESSAGES.PRODUCT_DELETED });
   } catch (error) {
